LayoutTests/http/tests/security/contentSecurityPolicy/javascript-url-blocked-expected.txt

CONSOLE MESSAGE: line 1: Refused to execute JavaScript URL because of Content-Security-Policy.
